requirements.txt
python-binance
pandas
matplotlib
websocket
numpy
plotly
